Move Delta to package
Helps with organization and in preparation for new Delta related interfaces
